fix bloom and add motion vectors

+++ BasicRenderer/include/Mesh/Mesh.h
@@ -25,8 +25,8 @@ public:
 	static std::shared_ptr<Mesh> CreateShared(std::unique_ptr<std::vector<std::byte>> vertices, unsigned int vertexSize, std::optional<std::unique_ptr<std::vector<std::byte>>> skinningVertices, unsigned int skinningVertexSize, const std::vector<UINT32>& indices, const std::shared_ptr<Material> material, unsigned int flags) {
 		return std::shared_ptr<Mesh>(new Mesh(std::move(vertices), vertexSize, std::move(skinningVertices), skinningVertexSize, indices, material, flags));
     }
-	uint32_t GetNumVertices(bool meshletReorderedVertices) const {
-		uint32_t size = meshletReorderedVertices ? m_meshletReorderedVertices.size() / m_perMeshBufferData.vertexByteSize : m_vertices->size() / m_perMeshBufferData.vertexByteSize;
+	uint64_t GetNumVertices(bool meshletReorderedVertices) const {
+		uint64_t size = meshletReorderedVertices ? m_meshletReorderedVertices.size() / m_perMeshBufferData.vertexByteSize : m_vertices->size() / m_perMeshBufferData.vertexByteSize;
 		return size;
 	}
     D3D12_VERTEX_BUFFER_VIEW GetVertexBufferView() const;
